requirements.txt
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
google-api-core==1.24.1
google-auth==1.24.0
google-cloud-speech==2.0.1
googleapis-common-protos==1.52.0
grpcio==1.34.0
idna==2.10
keyboard==0.13.5
libcst==0.3.16
mypy-extensions==0.4.3
ply==3.11
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
